Mudaca do tipo da variavel peso

+++ PastaFontes/Produto.java.java
@@ -3,19 +3,19 @@
 
 public class Produto {
 
-	private double peso;
+	private int peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(double peso, double altura) {
+	public Produto(int peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
